Add WordPress support.
The WordPress support is served when format = wpquicklatex. For example,

latex2markdown(sourc-filename,'format', 'wpquicklatex')

With this parameter, the latex2markdown() generate a MD source file which
is compatible with WordPress. The WordPress writer can paste this MD source
into the Gutenberg editor.

This generated MD assumes WP QuickLaTex plugin to render the LaTeX equations
( https://wordpress.org/plugins/wp-quicklatex/ ). The inline equation format
is $ ... $. And the display equation format is
\[
...\\
...\\
...
\]

In addition to above format, the [latex] short code is embeded at the
top of MD source file.

The figures are left as broken link. Thus, the WordPress writer can drag and
drop the figure files to the broken file marks of the Gutenberg editor.

Note, the Enlighter seems to  ignores the language mark of the code block.

Tested versions
WordPress     : 5.4.1-ja
WP QuickLaTeX : 3.8.6
Enlighter     : 4.2
Gutenberg     : 8.1.0